Bugfix: mtu size was 8 bits when this is required 16 bits for it

+++ src/coreL2/SystemParameters/StaticDefaultParameters.cpp
@@ -140,7 +140,7 @@ StaticDefaultParameters::StaticDefaultParameters(
 
 		//Gets Maximum Transmission Unity in Bytes
 		getline(defaultConfigurationsFile, readBuffer);
-		mtu[i] = stoi(readBuffer);
+		mtu[i] = stol(readBuffer);
 		readBuffer.clear();
 
 		//Gets IP Timeout
